Feat: Feed에 User Domain기능 추가

+++ src/main/kotlin/team/springpsring/petpartner/domain/feed/service/FeedService.kt
@@ -6,6 +6,7 @@ import org.springframework.transaction.annotation.Transactional
 import team.springpsring.petpartner.domain.feed.comment.entity.toResponse
 import team.springpsring.petpartner.domain.feed.comment.repository.CommentRepository
 import team.springpsring.petpartner.domain.feed.dto.CreateFeedRequest
+import team.springpsring.petpartner.domain.feed.dto.DeleteFeedRequest
 import team.springpsring.petpartner.domain.feed.dto.FeedResponse
 import team.springpsring.petpartner.domain.feed.dto.UpdateFeedRequest
 import team.springpsring.petpartner.domain.feed.entity.Feed
@@ -14,14 +15,26 @@ import team.springpsring.petpartner.domain.feed.repository.FeedRepository
 import team.springpsring.petpartner.domain.love.dto.CreateLoveRequest
 import team.springpsring.petpartner.domain.love.repository.LoveRepository
 import team.springpsring.petpartner.domain.love.service.LoveService
+import team.springpsring.petpartner.domain.user.dto.GetUserInfoRequest
+import team.springpsring.petpartner.domain.user.dto.UserResponse
+import team.springpsring.petpartner.domain.user.entity.User
+import team.springpsring.petpartner.domain.user.service.UserService
 
 @Service
 class FeedService(
     private val feedRepository: FeedRepository,
     private val commentRepository: CommentRepository,
     private val loveRepository: LoveRepository,
-    private val loveService: LoveService
+    private val loveService: LoveService,
+    private val userService: UserService
 ) {
+
+    @Transactional
+    fun checkValidate(token: String):UserResponse{
+        return userService.getUserInfo(
+            GetUserInfoRequest(
+                token=token))
+    }
 
     fun getAllFeeds(): List<FeedResponse> {
         return feedRepository.findAll().map { it.toResponse() }
@@ -45,23 +58,22 @@ class FeedService(
     }
 
     @Transactional
-    fun createFeed(createFeedRequest: CreateFeedRequest): FeedResponse {
+    fun createFeed(userInfo:UserResponse,createFeedRequest: CreateFeedRequest): FeedResponse {
         return feedRepository.save(
             Feed(
-                createFeedRequest.name,
+                userInfo.username,
                 createFeedRequest.title,
                 createFeedRequest.body,
                 createFeedRequest.images,
                 createFeedRequest.category,
-                0,
-                createFeedRequest.created
             )
         ).toResponse()
     }
 
     @Transactional
-    fun updateFeed(feedId: Long, updateFeedRequest: UpdateFeedRequest): FeedResponse {
+    fun updateFeed(userInfo:UserResponse ,feedId: Long, updateFeedRequest: UpdateFeedRequest): FeedResponse {
         val feed = feedRepository.findByIdOrNull(feedId) ?: throw NullPointerException("Feed not found")
+        if(feed.name!=userInfo.username) throw ArithmeticException("userNotMatch")
         val (title, body, images, category) = updateFeedRequest
 
         feed.title = title
@@ -73,19 +85,20 @@ class FeedService(
     }
 
     @Transactional
-    fun deleteFeed(feedId: Long) {
+    fun deleteFeed(userInfo:UserResponse ,feedId: Long) {
         val feed = feedRepository.findByIdOrNull(feedId) ?: throw NullPointerException("Feed not found")
+        if(feed.name!=userInfo.username) throw ArithmeticException("userNotMatch")
         feedRepository.delete(feed)
     }
 
     @Transactional
-    fun updateLoveForFeed(feedId: Long, isLove: Boolean, createLoveRequest: CreateLoveRequest) {
+    fun updateLoveForFeed(feedId: Long, isLove: Boolean, userInfo: UserResponse) {
         if(isLove){
-            val love = loveRepository.findByFeedIdAndLoginId(feedId,createLoveRequest.loginId)?:throw NullPointerException("Love not found")
+            val love = loveRepository.findByFeedIdAndLoginId(feedId,userInfo.loginId)?:throw NullPointerException("Love not found")
             loveService.deleteLove(love.id!!)
         }
         else{
-            loveService.createLove(feedId, createLoveRequest)
+            loveService.createLove(feedId, userInfo.loginId)
         }
     }
 }
